fix/checagem antes de deletar demoday e project

+++ src/main/java/com/demodayapi/controller/DemodayController.java
@@ -77,10 +77,9 @@ public class DemodayController {
 
 
      @DeleteMapping("/deletedemoday/{id}")
-        public ResponseEntity<Void> deleteDemoday(@PathVariable int id) {
-
-       demodayService.deleteDemodayById(id);
-            
+        public ResponseEntity<Void> deleteDemoday(@PathVariable int id,HttpServletRequest request) {
+        if(!userService.isLoggedUserAdmin(request))throw new UserIsNotAdminException();
+        demodayService.deleteDemodayById(id); 
         return ResponseEntity.noContent().build();
     }
